+++ client/src/types/CreateBillRequest.ts
@@ -1,0 +1,4 @@
+export interface CreateBillRequest {
+  name: string
+  billDate: string
+}
