--- client/src/types/CreateTodoRequest.ts
@@ -1,4 +1,0 @@
-export interface CreateTodoRequest {
-  name: string
-  dueDate: string
-}
