--- client/src/types/Todo.ts
@@ -1,8 +1,0 @@
-export interface Todo {
-  todoId: string
-  createdAt: string
-  name: string
-  dueDate: string
-  done: boolean
-  attachmentUrl?: string
-}
