+++ client/src/types/UpdateBillRequest.ts
@@ -1,6 +1,3 @@
-/**
- * Fields in a request to update a single TODO item.
- */
 export interface UpdateBillRequest {
   name: string
   billDate: string
