--- client/src/types/UpdateTodoRequest.ts
@@ -1,5 +1,0 @@
-export interface UpdateTodoRequest {
-  name: string
-  dueDate: string
-  done: boolean
-}
